+++ src/MonoTorrent.Client/MonoTorrent.Client/Unchokers/InitialSeedUnchoker.cs
@@ -206,7 +206,7 @@ namespace MonoTorrent.Client
                 data.TotalPieces++;
                 data.CurrentPieces[index] = true;
                 advertisedPieces.Add (new SeededPiece (data.Peer, index, Manager.Torrent!.PieceLength / Constants.BlockSize));
-                data.Peer.MessageQueue.Enqueue (BtEncoder.WriteHave (index));
+                data.Peer.MessageQueue.Enqueue (MessageEncoder.WriteHave (index));
                 index++;
             }
         }
